Fix bugs in EphemeralFileSystemAbstraction

+++ community/io/src/test/java/org/neo4j/graphdb/mockfs/EphemeralFileSystemAbstraction.java
@@ -47,6 +47,7 @@ import java.time.Clock;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
+import java.util.Comparator;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Iterator;
@@ -377,18 +378,22 @@ public class EphemeralFileSystemAbstraction implements FileSystemAbstraction
     }
 
     @Override
-    public void deleteRecursively( File directory ) throws IOException
+    public void deleteRecursively( File path ) throws IOException
     {
-        List<String> directoryPathItems = splitPath( canonicalFile( directory ) );
-        for ( Map.Entry<File,EphemeralFileData> file : files.entrySet() )
+        if ( isDirectory( path ) )
         {
-            File fileName = file.getKey();
-            List<String> fileNamePathItems = splitPath( fileName );
-            if ( directoryMatches( directoryPathItems, fileNamePathItems ) )
+            List<String> directoryPathItems = splitPath( canonicalFile( path ) );
+            for ( Map.Entry<File,EphemeralFileData> file : files.entrySet() )
             {
-                deleteFile( fileName );
+                File fileName = file.getKey();
+                List<String> fileNamePathItems = splitPath( fileName );
+                if ( directoryMatches( directoryPathItems, fileNamePathItems ) )
+                {
+                    deleteFile( fileName );
+                }
             }
         }
+        deleteFile( path );
     }
 
     @Override
@@ -437,7 +442,7 @@ public class EphemeralFileSystemAbstraction implements FileSystemAbstraction
             List<String> fileNamePathItems = splitPath( file );
             if ( directoryMatches( directoryPathItems, fileNamePathItems ) )
             {
-                found.add( constructPath( fileNamePathItems, directoryPathItems.size() + 1 ) );
+                found.add( constructPath( fileNamePathItems, directoryPathItems ) );
             }
         }
 
@@ -463,7 +468,7 @@ public class EphemeralFileSystemAbstraction implements FileSystemAbstraction
             List<String> fileNamePathItems = splitPath( file );
             if ( directoryMatches( directoryPathItems, fileNamePathItems ) )
             {
-                File path = constructPath( fileNamePathItems, directoryPathItems.size() + 1 );
+                File path = constructPath( fileNamePathItems, directoryPathItems );
                 if ( filter.accept( path.getParentFile(), path.getName() ) )
                 {
                     found.add( path );
@@ -473,10 +478,15 @@ public class EphemeralFileSystemAbstraction implements FileSystemAbstraction
         return found.toArray( new File[found.size()] );
     }
 
-    private File constructPath( List<String> pathItems, int count )
+    private File constructPath( List<String> pathItems, List<String> base )
     {
         File file = null;
-        for ( String pathItem : pathItems.subList( 0, count ) )
+        if ( base.size() > 0 )
+        {
+            // We're not directly basing off the root directory
+            pathItems = pathItems.subList( 0, base.size() + 1 );
+        }
+        for ( String pathItem : pathItems )
         {
             file = file == null ? new File( pathItem ) : new File( file, pathItem );
         }
@@ -497,12 +507,25 @@ public class EphemeralFileSystemAbstraction implements FileSystemAbstraction
     @Override
     public void moveToDirectory( File file, File toDirectory ) throws IOException
     {
-        EphemeralFileData fileToMove = files.remove( canonicalFile( file ) );
-        if ( fileToMove == null )
+        if ( isDirectory( file ) )
         {
-            throw new FileNotFoundException( file.getPath() );
+            File inner = new File( toDirectory, file.getName() );
+            mkdir( inner );
+            for ( File f : listFiles( file ) )
+            {
+                moveToDirectory( f, inner );
+            }
+            deleteFile( file );
         }
-        files.put( canonicalFile( new File( toDirectory, file.getName() ) ), fileToMove );
+        else
+        {
+            EphemeralFileData fileToMove = files.remove( canonicalFile( file ) );
+            if ( fileToMove == null )
+            {
+                throw new FileNotFoundException( file.getPath() );
+            }
+            files.put( canonicalFile( new File( toDirectory, file.getName() ) ), fileToMove );
+        }
     }
 
     @Override
@@ -546,7 +569,7 @@ public class EphemeralFileSystemAbstraction implements FileSystemAbstraction
         List<File> names = new ArrayList<>( files.size() );
         names.addAll( files.keySet() );
 
-        Collections.sort( names, ( o1, o2 ) -> o1.getAbsolutePath().compareTo( o2.getAbsolutePath() ) );
+        names.sort( Comparator.comparing( File::getAbsolutePath ) );
 
         for ( File name : names )
         {
@@ -888,14 +911,7 @@ public class EphemeralFileSystemAbstraction implements FileSystemAbstraction
 
     private static class EphemeralFileData
     {
-        private static final ThreadLocal<byte[]> SCRATCH_PAD = new ThreadLocal<byte[]>()
-        {
-            @Override
-            protected byte[] initialValue()
-            {
-                return new byte[1024];
-            }
-        };
+        private static final ThreadLocal<byte[]> SCRATCH_PAD = ThreadLocal.withInitial( () -> new byte[1024] );
         private DynamicByteBuffer fileAsBuffer;
         private DynamicByteBuffer forcedBuffer;
         private final Collection<WeakReference<EphemeralFileChannel>> channels = new LinkedList<>();
